+++ test/RuleMintAllowance/RuleMintAllowanceOwnable2Step.t.sol
@@ -9,6 +9,8 @@ import {ERC1404ExtendInterfaceId} from "CMTAT/library/ERC1404ExtendInterfaceId.s
 import {RuleEngineInterfaceId} from "CMTAT/library/RuleEngineInterfaceId.sol";
 import {OwnableInterfaceId} from "RuleEngine/modules/library/OwnableInterfaceId.sol";
 import {Ownable2StepInterfaceId} from "RuleEngine/modules/library/Ownable2StepInterfaceId.sol";
+import {IERC7551Compliance} from "CMTAT/interfaces/tokenization/draft-IERC7551.sol";
+import {IERC3643ComplianceFull} from "src/mocks/IERC3643ComplianceFull.sol";
 
 contract RuleMintAllowanceOwnable2StepTest is Test, HelperContract {
     address constant OWNER = address(1);
@@ -94,6 +96,8 @@ contract RuleMintAllowanceOwnable2StepTest is Test, HelperContract {
         assertTrue(rule.supportsInterface(RuleInterfaceId.IRULE_INTERFACE_ID));
         assertTrue(rule.supportsInterface(ERC1404ExtendInterfaceId.ERC1404EXTEND_INTERFACE_ID));
         assertTrue(rule.supportsInterface(RuleEngineInterfaceId.RULE_ENGINE_INTERFACE_ID));
+        assertTrue(rule.supportsInterface(type(IERC7551Compliance).interfaceId));
+        assertTrue(rule.supportsInterface(type(IERC3643ComplianceFull).interfaceId));
         assertFalse(rule.supportsInterface(bytes4(0xdeadbeef)));
     }
 }